Event Management Architecture
- Moved eventTypes up a level
- Create EventFeed as root component
- Pass selected event type up on create click
