+++ client/EventFeed/EventFeed.js
@@ -1,4 +1,19 @@
 import React from 'react';
+import EventCreator from '../EventCreator/EventCreator.js';
+
+let eventTypes = [{
+  id: '0',
+  name: 'encounter',
+  label: 'Pokemon Encounter'
+}, {
+  id: '1',
+  name: 'pokemonStatusUpdate',
+  label: 'Pokemon Update'
+}, {
+  id: '2',
+  name: 'goalStatusUpdate',
+  label: 'Goal Update'
+}];
 
 export default class EventFeed extends React.Component {
   constructor(props){
@@ -8,8 +23,15 @@ export default class EventFeed extends React.Component {
   render(){
     return (
       <div className="eventFeed">
-        
+        <EventCreator eventTypes={eventTypes} createEvent={this.createEvent}/>
       </div>
     )
+  }
+
+  /**
+   * Exposed Actions
+   */
+  createEvent(newEvent){
+    console.log(`Creating Event Type: ${newEvent.type}`);
   }
 }
